+++ src/gen_why3.ml
@@ -1781,6 +1781,8 @@ let rec map_mterm m ctx (mt : M.mterm) : loc_term =
     | Mdiveuc (l, r) -> Tdiv (dl Tyint, map_mterm m ctx l, map_mterm m ctx r)
     | Mmodulo (l, r) -> Tmod (dl Tyint, map_mterm m ctx l, map_mterm m ctx r)
     | Muminus v -> Tuminus (dl Tyint, map_mterm m ctx v)
+    | Mshiftleft (_l, _r)  -> error_not_translated "Mshiftleft"
+    | Mshiftright (_l, _r) -> error_not_translated "Mshiftright"
 
 
     (* asset api effect *)